Task 1707
(Incorporate attribute metadata into search)
https://rm.mventory.com/issues/1707

- AbsProductActivity: removed nameV custom on long click listener
initialization from the absOnCreate method. This functionality is moved
to initSpecialAttributes method
- AbsProductActivity: added initSpecialAttributes method
- AbsProductActivity: added initSpecialAttributes call to the
onAttributeListLoadSuccess method
- AbsProductActivity: added popup menu initialization to the
initDescriptionField method. Popup menu replaced context menu
- AbsProductActivity: removed onCreateContextMenu, onContextItemSelected
methods custom implementation
- AbsProductActivity: removed permanent product name passing to the
searchCriteriaParts in the startWebActivity method
- AbsProductActivity: removed permanent product name and description
attributes passing in the startWebActivity method
- AbsProductActivity: added conditional passing of name and description
parameters to the searchCriteriaParts and textAttributes in the
startWebActivity method
- ProductDetailsActivity: added skipping of product special attributes
in the mapData method when iterating through attributes list
- ProductDetailsActivity: remed permanent name attribute passing to the
search criteria parts in the startWebActivity. Now that attribute is
present in the mCustomAttributes so it will be handled there as simple
attribute
- JobCacheManager: incremented CACHE_VERSION constant to clear the cache
on first start of this app version
- CustomAttributesList: added mSpecialCustomAttributes field and its
getter
- CustomAttributesList: incremented serialVersionUID constant
- CustomAttributesList: added mSpecialCustomAttributes field
initialization to the loadFromAttributeList method
- ProductAttributeFullInfoProcessor: removed SPECIAL_ATTRIBUTES
filtering from the process method. Not it is handled in other places
- res/values/styles_theme.xml: added PopupMenuTextLarge.Theme style
declaration
- res/values/themes_theme.xml: added
android:textAppearanceLargePopupMenu custom attribute value to the
_Theme declaration